Restructure developer documentation
I reordered the menu elements according the decision of the core developers meeting:
The section "getting started" is kept alive with the sub menu point "installation". The section "update" goes to the last place as someone will have to deal with a update after installing or extending a OXID eShop in an ordinary case. The structure should be fixed in order to link to this documentation. But the menu entries can be renamed at some point in the future.
Additionally I restructured the section "modules" with a new section "skeleton". This section contains necessary information about the structure and metadata files in order to implement a module properly. The section "modules/developing" was dropped because all topics under the section "modules" are more or less about "developing".