yml is and must be yaml instead
The OXID eShop 6.2 generates and uses only the full file extension .yaml and not just .yml. If you change the configuration files to .yml, the module section at the administration area crashes and the shop becomes unstable. So it is a real issue and the documentation has to insist on yaml instead of yml to prevent misconfiguration.

Unfortunately this issue is possibly present at many other sections in the documentation and should be verified by the maintainer.