6.1 is only compatible with Composer 1
The OXID eShop version 6.1 is only compatible with Composer 1. If you upgrade to the latest version (2.0) as suggested before, installation will fail.

This issue was corrected in the user docs.
https://docs.oxid-esales.com/eshop/en/6.1/installation/new-installation/server-and-system-requirements.html#composer

The 6.2 developer docs simply refer to the user docs:
https://docs.oxid-esales.com/developer/en/6.2/getting_started/installation/environment_preparation.html#checking-system-compatibility

I would suggest to do the same with version 6.1 to have a single point of information. Otherwise you need to carefully verify such changes in both docs, user and developer.