Removed command name in services.yaml
While omitting the command name in the services.yaml file could lead to errors in older shop versions, it's not an issue anymore. We omit this part for quite some time in Academy trainings and had never problems. This goes back some version 6 iteration but to make it easy, I suggest to don't change any v6 documentation. Nevertheless we should omit it starting with OXID eShop 7.0 documentation.

The issue was due to an older symfony/console version (<3.4) used by Composer that missed a method. However, current shop 7.0 uses symfony/console 6.2 and composer 2.4 which supports "^5.4.11 || ^6.0.11". Therefore definitely not a problem anymore.